crawler: fix awfull mistake
the #hastag_tweets file is appended on every run, while it
should be rewritten. Due to this, the counted tweets are
way to high. Correct this.

While at it, rename the print file to printer to avoid any
unwanted errors while "print" is a default python command
and it should not be used for other things

Also post the desisered info and graphs to twitter

Signed-off-by: Erik Castricum <git@cas-online.nl>